requirements.txt
feedparser>=6.0.0
aiohttp>=3.9.0
openai>=1.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.0
fake-useragent>=1.4.0

# 可选：用于绕过顽固付费墙（Bloomberg, FT, Washington Post）
# 安装后需运行: playwright install chromium
playwright>=1.40.0

# Telegram 机器人
python-telegram-bot>=22.6

pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=5.0.0
pytest-xdist>=3.5.0
coverage>=7.4.0
httpx>=0.26.0
responses>=0.24.0
freezegun>=1.2.0
pytest-dotenv>=0.5.0
